pyproject.toml
[tool.poetry]
name = "test-tiff"
version = "0.1.0"
description = ""
authors = ["TurtleOld <dev@pavlovteam.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
